reuse SourcegraphCompletionsClient.complete across browser/node
They both use the standard `fetch` API, so the impl can be shared.
